OCSADV-200: move new sync action to same package as all the other actions
