+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/viewer/SPViewerMenuBar.java
@@ -81,7 +81,7 @@ final class SPViewerMenuBar extends JMenuBar {
         menu.addSeparator();
         menu.add(new ImportAction(_viewer));
         menu.add(new ExportAction(_viewer));
-        menu.add(_viewer._actions.vcs2SyncAction);
+        menu.add(_viewer._actions.vcsSyncAction);
         menu.add(_viewer._actions.syncAllAction);
         menu.addSeparator();
         menu.add(new FetchLibrariesAction(_viewer));
